fix: clamp incident distance to route length and align type with point events
- Change distanceFromStartMeters type from Int to Double to match RoutePointEvent
- Apply coerceIn(0.0, totalGeometryMetres) upper bound clamp in ExtNavRouteIncidentMapper
- Update visibleRouteIncidents filter to use Double comparison without toInt() cast
- Add test case verifying distances beyond route length are clamped to totalGeometryMetres
- Update existing test assertions from Int to Double literals

+++ core/model/src/androidMain/kotlin/me/matsumo/onenavi/core/model/RouteIncidentMarker.kt
@@ -29,7 +29,7 @@ data class RouteIncidentMarker(
     val category: RouteIncidentMarkerCategory,
     val coord: RoutePoint,
     val displayText: String,
-    val distanceFromStartMeters: Int,
+    val distanceFromStartMeters: Double,
     val polylinePointIndex: Int,
     val placeName: String?,
     val roadNumbering: String?,
